# api/NeoWallet_OpenAPI_v1.yaml
openapi: 3.1.0
info:
  title: NeoWallet API
  version: 1.0.0
  description: NeoWallet MVP API Contract Specification
  contact:
    name: NeoWallet API Team
    email: api@neowallet.com
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: https://api.neowallet.com/api/v1
    description: Production
  - url: https://api-staging.neowallet.com/api/v1
    description: Staging
  - url: https://api-dev.neowallet.com/api/v1
    description: Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - requestId
        - timestamp
      properties:
        code:
          type: string
          example: BUDGET_NOT_FOUND
        message:
          type: string
          example: Budget was not found.
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        totalCount:
          type: integer
        pageCount:
          type: integer

    # Authentication Schemas
...
    SavingsGoal:
      type: object
      properties:
        goalId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          maxLength: 100
        targetAmount:
          type: number
          format: decimal
          exclusiveMinimum: 0
        targetDate:
          type: string
          format: date
        currentAmount:
          type: number
          format: decimal
          minimum: 0
        currency:
          type: string
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        category:
          type: string
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELLED]
        progressPercentage:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSavingsGoalRequest:
      type: object
      required:
        - name
        - targetAmount
        - targetDate
      properties:
        name:
          type: string
          maxLength: 100
        targetAmount:
          type: number
          format: decimal
          exclusiveMinimum: 0
        targetDate:
          type: string
          format: date
        currentAmount:
          type: number
          format: decimal
          minimum: 0
          default: 0
        currency:
          type: string
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
          default: MEDIUM
        category:
          type: string
    UpdateSavingsGoalRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        targetAmount:
          type: number
          format: decimal
          exclusiveMinimum: 0
        targetDate:
          type: string
          format: date
        currentAmount:
          type: number
          format: decimal
          minimum: 0
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        category:
          type: string
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELLED]
    SavingsGoalProgress:
      type: object
      properties:
        goalId:
          type: string
          format: uuid
        currentAmount:
          type: number
          format: decimal
        targetAmount:
          type: number
          format: decimal
        progressPercentage:
          type: number
          format: decimal
        remainingAmount:
          type: number
          format: decimal
        monthsRemaining:
          type: integer
        requiredMonthlyContribution:
          type: number
          format: decimal
        actualMonthlyContribution:
          type: number
          format: decimal
        onTrack:
          type: boolean
        status:
          type: string
          enum: [ON_TRACK, BEHIND, AHEAD]
        calculatedAt:
          type: string
          format: date-time
    SavingsGoalForecast:
      type: object
      properties:
        goalId:
          type: string
          format: uuid
        forecast:
          type: object
          properties:
            projectedCompletionDate:
              type: string
              format: date
            projectedCompletionAmount:
              type: number
              format: decimal
            monthsToCompletion:
              type: integer
            confidence:
              type: string
              enum: [HIGH, MEDIUM, LOW]
        scenarios:
          type: array
          items:
            type: object
            properties:
              monthlyContribution:
                type: number
                format: decimal
              projectedCompletionDate:
                type: string
                format: date
              monthsToCompletion:
                type: integer
        calculatedAt:
          type: string
          format: date-time
    CreateSavingsContributionRequest:
      type: object
      required:
        - amount
        - contributionDate
      properties:
        amount:
          type: number
          format: decimal
          exclusiveMinimum: 0
        contributionDate:
          type: string
          format: date
        notes:
          type: string
          maxLength: 500
    SavingsContribution:
      type: object
      properties:
        contributionId:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        currency:
          type: string
        contributionDate:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    SavingsContributionResult:
      type: object
      properties:
        contribution:
          $ref: '#/components/schemas/SavingsContribution'
        goal:
          $ref: '#/components/schemas/SavingsGoal'

    # Bill Schemas
...
paths:
...
  /savings-goals:
    post:
      summary: Create savings goal
      operationId: createSavingsGoal
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: X-Family-ID
          in: header
          required: false
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSavingsGoalRequest'
      responses:
        '201':
          description: Savings goal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get savings goals
      operationId: getSavingsGoals
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: X-Family-ID
          in: header
          required: false
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: filter[status]
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, CANCELLED]
        - name: filter[priority]
          in: query
          schema:
            type: string
            enum: [HIGH, MEDIUM, LOW]
        - name: sort
          in: query
          schema:
            type: string
            example: targetDate:asc
      responses:
        '200':
          description: Savings goals retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  goals:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavingsGoal'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /savings-goals/{goalId}:
    get:
      summary: Get savings goal by ID
      operationId: getSavingsGoal
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Savings goal retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update savings goal
      operationId: updateSavingsGoal
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSavingsGoalRequest'
      responses:
        '200':
          description: Savings goal updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete savings goal
      operationId: deleteSavingsGoal
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Savings goal deleted successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /savings-goals/{goalId}/progress:
    get:
      summary: Get savings goal progress
      operationId: getSavingsGoalProgress
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Savings goal progress retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalProgress'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /savings-goals/{goalId}/forecast:
    get:
      summary: Get savings goal forecast
      operationId: getSavingsGoalForecast
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Savings goal forecast retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalForecast'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /savings-goals/{goalId}/contributions:
    post:
      summary: Add contribution to a savings goal
      operationId: addSavingsGoalContribution
      tags:
        - Savings Goals
      security:
        - bearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Family-ID
          in: header
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSavingsContributionRequest'
      responses:
        '201':
          description: Contribution recorded and goal progress updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsContributionResult'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Bill Paths
...